requirements.txt
asgiref==3.8.1
astroid==3.2.4
crispy-bootstrap5==2024.2
dill==0.3.8
Django==4.2.15
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.2
pylint==3.2.6
python-decouple==3.8
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing-extensions==4.12.2
